requirements.txt
flake8==7.0.0
ipykernel==6.29.3
ipywidgets==8.1.2
markdown
python-dotenv>=1.0.1
jupyter>=1.0.0

backoff==2.2.1
openai>=1.35.4
langchain
langchain_community
numpy
PyPDF2>=3.0.0

# Python >=3.7
# poppler-utils is required for pdf2image
pdf2image==1.17.0
# Python >=3.8
requests>=2.23.2
# Python >=3.8
boto3>=1.34.144

marker-pdf
ocrmypdf
